web/dumb_login/challenge/requirements.txt
Flask==2.2.5
Pillow==10.1.0
pytesseract==0.3.10
